Define frame data structure
Define the data structure in which we will represent an OCR frame and
document it.

Eventually we will want to duplicate this documentation outside of
source code or move it outside source entirely into a wiki or
something. For now, while prototyping/doing a first pass on the
feature, it is simplest and best to leave it in source.