isTwo() passes tests
Wrote isTwo() method and passes tests.
NOTE: We have some junk data for THREE, FOUR, ETC right now in the repo.
But there are no tests or comparison functions so it is still easy to
see application state and the application is still in a working state.

Iteratively making the valid-frames data real one number at a time and
adding comparison function and tests.